refactor: de-part naval_units_panel cluster (#4117)
Migrate the 5-part naval units panel to explicit-import mixin libraries
with a public NavalUnitsPanelState; shrink no-part grandfather by 6.

Refs #4117
