+++ app/lib/features/game/widgets/units/naval/naval_units_panel_support_dialogs.dart
@@ -1,14 +1,29 @@
-/// Dialog openers for naval fleet actions. SPEC/ui/naval-units-panel.md.
+import 'package:colonizethis_logic/colonizethis_logic.dart' show GamePlayerLookup;
+import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:flutter/material.dart';
 
-part of 'naval_units_panel.dart';
+import '../../../../../core/services/app_event_bus_panel_nav.dart';
+import '../../../../../core/services/app_event_handler/app_event_handler_scope.dart'
+    show trainNavalDialogId;
+import '../../panels/tree_builders/naval_tree_builder.dart';
+import '../../unit_orders/move_fleet_dialog.dart';
+import '../../unit_orders/split_fleet_dialog.dart';
+import '../shared/base_units_panel.dart';
+import 'naval_units_panel.dart';
+import 'naval_units_panel_state_base.dart';
+import 'naval_units_panel_support_combine.dart';
 
-extension _NavalUnitsPanelDialogs on _NavalUnitsPanelState {
-  void _openTrainDialog() {
+mixin NavalUnitsPanelDialogs
+    on
+        BaseUnitsPanelState<NavalUnitsPanel>,
+        NavalUnitsPanelStateBase,
+        NavalUnitsPanelCombine {
+  void openTrainDialog() {
     widget.bus.closePanelThenEmit(OpenDialogEvent(trainNavalDialogId));
   }
 
-  void _openSplitDialog(FleetRow row) {
-    final id = _selectionFleetId(row);
+  void openSplitDialog(FleetRow row) {
+    final id = selectionFleetId(row);
     final fleet = widget.game.fleetById(id);
     if (fleet == null) return;
 
@@ -25,7 +40,7 @@ extension _NavalUnitsPanelDialogs on _NavalUnitsPanelState {
     );
   }
 
-  Future<void> _openMoveFleetDialog(FleetRow row) async {
+  Future<void> openMoveFleetDialog(FleetRow row) async {
     if (row.isHomeFleet) return;
     final fleet = widget.game.fleetById(row.fleetId);
     final nonNullFleet = fleet;
