+++ tool/app_no_part_directives_grandfather.dart
@@ -1,4 +1,4 @@
-// Generated grandfather allowlist: 182 app/lib files still using part directives.
+// Generated grandfather allowlist: 176 app/lib files still using part directives.
 // Shrink-only: remove entries as de-part slices land (Refs #4117).
 const List<String> appNoPartDirectivesGrandfatheredForTests = <String>[
   'app/lib/core/services/debug/app_event_handler_debug_flip_province.dart',
@@ -141,12 +141,6 @@ const List<String> appNoPartDirectivesGrandfatheredForTests = <String>[
   'app/lib/features/game/widgets/units/military/military_units_panel_dialogs.dart',
   'app/lib/features/game/widgets/units/military/military_units_panel_support_army_tile.dart',
   'app/lib/features/game/widgets/units/military/military_units_panel_support_detail_rows.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel_build.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel_list.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel_support_combine.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel_support_dialogs.dart',
-  'app/lib/features/game/widgets/units/naval/naval_units_panel_support_home_transfer.dart',
   'app/lib/features/game/widgets/units/shared/units_entity_action_row.dart',
   'app/lib/features/game/widgets/units/shared/units_entity_action_row_actions.dart',
   'app/lib/features/shell/new_game_setup_flow.dart',
